fix: 실시간 선수 성적 불러오기 기능 DTO 생성 및 VO 수정 #16
- LiveScoreOfTeamWithPlayerVO
- LiveScoreOfTeam

+++ be/src/main/java/kr/codesquad/baseball/dto/playerVO/BatterLiveScoreVO.java
@@ -18,7 +18,7 @@ public class BatterLiveScoreVO extends PlayerVO {
     private int outCount;
 
     @Builder
-    public BatterLiveScoreVO(int playerId, String playerName, int battingOrder, int battingAverage, int plateAppearance, int hitCount, int outCount) {
+    public BatterLiveScoreVO(int playerId, String playerName, int battingOrder, double battingAverage, int plateAppearance, int hitCount, int outCount) {
         super(playerId, playerName);
         this.battingOrder = battingOrder;
         this.battingAverage = battingAverage;
